backend/requirements.txt
requests==2.31.0
feedparser==6.0.10
beautifulsoup4==4.12.2
python-dateutil==2.8.2
lxml==4.9.3
Flask==2.3.3
gunicorn==21.2.0
# Optional: For advanced ML features (uncomment if needed)
# scikit-learn==1.3.0
# nltk==3.8.1 